requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psutil==5.9.6
websockets==12.0
Pillow==10.0.1
pystray==0.22.0
win10toast==0.9
